Lift the automation endpoints out of web_server
api/automations.py takes the /api/automations family - list, master
pause, create/update/group/bulk-toggle, delete, duplicate, toggle, run,
progress, history, blocks, test-notify - plus /api/scripts for the
run-script block dropdown. the route bodies were already thin handlers
over core/automation/api.py; now the thin layer lives out here too.

bodies byte-identical; only the decorator changed. automation_engine is
injected as the object (bound once at boot, never rebound; the wiring
runs after that binding). routemap identical on method+path, automation
suite green (405), video renderer contract green.
web_server: 36,526 -> 36,286 lines.
